+++ CHANGES.md
@@ -11,6 +11,7 @@ Improvements 🙌:
  - Hide left rooms in breadcrumbs (#766)
  - Correctly handle SSO login redirection
  - SSO login is now performed in the default browser, or in Chrome Custom tab if available (#1400)
+ - Improve checking of homeserver version support (#1442)
 
 Bugfix 🐛:
  - Switch theme is not fully taken into account without restarting the app